Issue #4014: Split and Organize Checkstyle inputs by Test for MethodCount
